requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
cssselect==1.2.0
cssutils==2.9.0
idna==3.6
lxml==5.1.0
premailer==3.10.0
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
urllib3==2.2.0
yagmail==0.15.293
